feat: Cidade atual do usuário é guardada com shared_preferences, garantindo que ela permaneça ou seja atualizada por ele ao decorrer da aplicação.
This commit adds two new variables, currentPageIndex and currentFilterCriteria, to the MyHomePageController class. These variables will be used to keep track of the current page index and filter criteria selected by the user. The currentPageIndex variable is of type RxInt and will be observed for changes, while the currentFilterCriteria variable is of type RxString. This commit also includes the necessary methods to update and load the selected city using shared preferences.

Ref: #5